requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.1
Django==3.0.3
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
django-extensions==2.2.8
django-restframework==0.0.1
django-tastypie==0.14.3
entrypoints==0.3
flake8==3.7.9
idna==2.9
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
parso==0.6.1
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.3
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
python-mimeparse==1.6.0
pytz==2019.3
requests==2.23.0
rope==0.16.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==1.9.5
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
wrapt==1.11.2
yapf==0.29.0
